+++ src/client/vite.config.js
@@ -18,6 +18,11 @@ export default defineConfig({
         target: "http://localhost:5000/",
         changeOrigin: true,
       },
+      "/dataextraction": {
+        target: "http://localhost:8000/",
+        changeOrigin: true,
+        rewrite: path => path.replace(/^\/dataextraction/, ""),
+      },
     },
     port: 3000,
   },